requirements.txt
django==3.1.7
djangorestframework==3.12.2
markdown==3.3.4
django-filter==2.4.0
python-dotenv==0.15.0
psycopg2-binary==2.8.6
celery==5.0.5
requests==2.25.1
django-celery-beat==2.2.0